Move love types to types/
