--- typings/love/structs.d.ts
@@ -1,2 +1,0 @@
-import "./Types";
-import "./UtilityTypes";
